Create models and some associations
