+++ app/models/like.rb
@@ -1,0 +1,4 @@
+class Like < ApplicationRecord
+  belongs_to :user, class_name: 'User'
+  belongs_to :post
+end
