+++ app/models/user.rb
@@ -1,0 +1,5 @@
+class User < ApplicationRecord
+    has_many :posts, dependent: :destroy
+    has_many :comments, dependent: :destroy
+    has_many :likes, dependent: :destroy
+end
