--- db/migrate/20231031100905_create_posts.rb
@@ -1,12 +1,0 @@
-class CreatePosts < ActiveRecord::Migration[7.1]
-  def change
-    create_table :posts do |t|
-t.string :title
-t.text :text
-t.integer :comments_counter
-t.integer :likes_counter
-t.references :author, null: false, foreign_key: {to_table: :users}
-      t.timestamps
-    end
-  end
-end
