+++ db/migrate/20231031204201_create_users.rb
@@ -1,10 +1,10 @@
 class CreateUsers < ActiveRecord::Migration[7.1]
   def change
     create_table :users do |t|
-t.string :name
-t.string :photo_link
-t.string :bio
-t.integer :posts_counter
+      t.string :name
+      t.string :photo_link
+      t.string :bio
+      t.integer :posts_counter
 
       t.timestamps
     end
