+++ db/migrate/20231031204951_create_posts.rb
@@ -1,0 +1,13 @@
+class CreatePosts < ActiveRecord::Migration[7.1]
+  def change
+    create_table :posts do |t|
+      t.string :title
+      t.text :text
+      t.integer :comments_counter
+      t.integer :likes_counter
+      t.references :author, null: false, foreign_key: { to_table: :users }
+
+      t.timestamps
+    end
+  end
+end
