+++ db/migrate/20231031205249_create_comments.rb
@@ -2,7 +2,7 @@ class CreateComments < ActiveRecord::Migration[7.1]
   def change
     create_table :comments do |t|
       t.text :text
-      t.references :user, null: false, foreign_key: { to_table: :users }
+      t.references :user, null: false, foreign_key: {to_table: :users}
       t.references :post, null: false, foreign_key: true
 
       t.timestamps
