+++ db/migrate/20231031205704_create_likes.rb
@@ -1,7 +1,7 @@
 class CreateLikes < ActiveRecord::Migration[7.1]
   def change
     create_table :likes do |t|
-      t.references :user, null: false, foreign_key: { to_table: :users }
+      t.references :user, null: false, foreign_key: {to_table: :users}
       t.references :post, null: false, foreign_key: true
 
       t.timestamps
