+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2023_10_31_101633) do
+ActiveRecord::Schema[7.1].define(version: 2023_10_31_205704) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
