+++ test/fixtures/posts.yml
@@ -1,0 +1,15 @@
+# Read about fixtures at https://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
+
+one:
+  title: MyString
+  text: MyText
+  comments_counter: 1
+  likes_counter: 1
+  author: one
+
+two:
+  title: MyString
+  text: MyText
+  comments_counter: 1
+  likes_counter: 1
+  author: two
