+++ test/fixtures/users.yml
@@ -1,0 +1,13 @@
+# Read about fixtures at https://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
+
+one:
+  name: MyString
+  photo_link: MyString
+  bio: MyString
+  posts_counter: 1
+
+two:
+  name: MyString
+  photo_link: MyString
+  bio: MyString
+  posts_counter: 1
