+++ test/models/like_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class LikeTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
